+++ winit/src/application.rs
@@ -861,6 +861,9 @@ pub fn run_command<A, C, E>(
                     .send_event(tagger(Ok(())))
                     .expect("Send message to event loop");
             }
+            command::Action::Custom(_) => {
+                log::warn!("Unsupported custom action in `iced_winit` shell");
+            }
         }
     }
 }