+++ CHANGELOG.md
@@ -31,6 +31,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `style` attribute for `Font`. [#2041](https://github.com/iced-rs/iced/pull/2041)
 - Texture filtering options for `Image`. [#1894](https://github.com/iced-rs/iced/pull/1894)
 - `default` and `shift_step` methods for `slider` widgets. [#2100](https://github.com/iced-rs/iced/pull/2100)
+- `Custom` variant to `command::Action`. [#2146](https://github.com/iced-rs/iced/pull/2146)
 
 ### Changed
 - Enable WebGPU backend in `wgpu` by default instead of WebGL. [#2068](https://github.com/iced-rs/iced/pull/2068)
@@ -101,6 +102,7 @@ Many thanks to...
 - @casperstorm
 - @cfrenette
 - @Davidster
+- @Decodetalkers
 - @derezzedex
 - @dtzxporter
 - @GyulyVGC